// src/types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.15"
  }
  public: {
// ...
    Enums: {
      growth_status: "Cultivated" | "Wild"
      life_form: "Annual" | "Perennial"
      map_layer: "trees" | "shrubs" | "herbs"
      native_status: "Native" | "Exotic"
      plant_type:
        | "Tree"
        | "Palm"
        | "Shrub"
        | "Subshrub"
        | "Climber"
        | "Succulent"
        | "Herb"
        | "Grass"
        | "Sedge"
      submission_status:
        | "Draft"
        | "Pending"
        | "Review"
        | "Approved"
        | "Rejected"
      submission_type:
        | "Photo"
        | "Observation"
        | "Specimen"
        | "GPS Pin"
        | "Photo + GPS"
        | "Draft"
      user_role: "admin" | "contributor" | "student"
    }
    CompositeTypes: {
      geometry_dump: {
        path: number[] | null
        geom: unknown
      }
      valid_detail: {
        valid: boolean | null
        reason: string | null
        location: unknown
      }
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never